+++ constant.ts
@@ -1,0 +1,1 @@
+export const API_HOST = 'http://3.35.229.111:8080';
